# packages/api-types/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KnowledgeBaseDev API",
    "version": "0.1.0-alpha"
  },
  "paths": {
...
    "/api/text-imports": {
      "post": {
        "summary": "Create Text Import",
        "operationId": "create_text_import_api_text_imports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/events:stream": {
      "get": {
        "summary": "Stream Job Events",
        "operationId": "stream_job_events_api_jobs__job_id__events_stream_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review-tasks": {
      "get": {
        "summary": "List Review Tasks",
        "operationId": "list_review_tasks_api_review_tasks_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pending_review",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewTask"
                  },
                  "title": "Response List Review Tasks Api Review Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review-tasks/{task_id}:confirm": {
      "post": {
        "summary": "Confirm Review Task",
        "operationId": "confirm_review_task_api_review_tasks__task_id__confirm_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review-tasks/{task_id}:ignore": {
      "post": {
        "summary": "Ignore Review Task",
        "operationId": "ignore_review_task_api_review_tasks__task_id__ignore_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/retrieval/evidence-only": {
      "post": {
        "summary": "Evidence Only",
        "operationId": "evidence_only_api_retrieval_evidence_only_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceOnlyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceOnlyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/retrieval/preview": {
      "post": {
        "summary": "Retrieval Preview",
        "operationId": "retrieval_preview_api_retrieval_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence-packs/{evidence_pack_id}": {
      "get": {
        "summary": "Evidence Pack Detail",
        "operationId": "evidence_pack_detail_api_evidence_packs__evidence_pack_id__get",
        "parameters": [
          {
            "name": "evidence_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Pack Id"
            }
          },
          {
            "name": "focus_item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Focus Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePackDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence-packs/{evidence_pack_id}/annotations": {
      "get": {
        "summary": "Evidence Pack Annotations",
        "operationId": "evidence_pack_annotations_api_evidence_packs__evidence_pack_id__annotations_get",
        "parameters": [
          {
            "name": "evidence_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationAnnotationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Evidence Pack Annotation",
        "operationId": "create_evidence_pack_annotation_api_evidence_packs__evidence_pack_id__annotations_post",
        "parameters": [
          {
            "name": "evidence_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Pack Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitationAnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationAnnotationRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/citation-annotations/{annotation_id}": {
      "patch": {
        "summary": "Patch Citation Annotation",
        "operationId": "patch_citation_annotation_api_citation_annotations__annotation_id__patch",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Annotation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitationAnnotationPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationAnnotationRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove Citation Annotation",
        "operationId": "remove_citation_annotation_api_citation_annotations__annotation_id__delete",
        "parameters": [
          {
            "name": "annotation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Annotation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Remove Citation Annotation Api Citation Annotations  Annotation Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence-packs/{evidence_pack_id}/compare": {
      "post": {
        "summary": "Compare Evidence Pack Items",
        "operationId": "compare_evidence_pack_items_api_evidence_packs__evidence_pack_id__compare_post",
        "parameters": [
          {
            "name": "evidence_pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Pack Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CitationCompareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationCompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CitationAnnotationListResponse": {
        "properties": {
          "evidence_pack_id": {
            "type": "string",
            "title": "Evidence Pack Id"
          },
          "annotations": {
            "items": {
              "$ref": "#/components/schemas/CitationAnnotationRecord"
            },
            "type": "array",
            "title": "Annotations"
          },
          "counts_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts By Type"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "evidence_pack_id",
          "annotations",
          "counts_by_type",
          "total"
        ],
        "title": "CitationAnnotationListResponse"
      },
...
      "EvidenceOnlyRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": "default-space"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "tag_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tag Ids"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "EvidenceOnlyRequest"
      },
      "EvidenceOnlyResponse": {
        "properties": {
          "retrieval_log_id": {
            "type": "string",
            "title": "Retrieval Log Id"
          },
          "evidence_pack_id": {
            "type": "string",
            "title": "Evidence Pack Id"
          },
          "answer_id": {
            "type": "string",
            "title": "Answer Id"
          },
          "output_type": {
            "type": "string",
            "title": "Output Type"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "evidence_item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Item Ids"
          },
          "citation_labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citation Labels"
          },
          "citation_trace_summary": {
            "type": "string",
            "title": "Citation Trace Summary"
          },
          "provider_status": {
            "type": "string",
            "title": "Provider Status"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          }
        },
        "type": "object",
        "required": [
          "retrieval_log_id",
          "evidence_pack_id",
          "answer_id",
          "output_type",
          "answer",
          "evidence_item_ids",
          "citation_labels",
          "citation_trace_summary",
          "provider_status",
          "fallback_reason"
        ],
        "title": "EvidenceOnlyResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobSnapshot": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/JobEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status",
          "trace_id",
          "payload",
          "result",
          "events"
        ],
        "title": "JobSnapshot"
      },
...
      "RetrievalPreviewRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": "default-space"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "tag_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tag Ids"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrievalPreviewRequest"
      },
...
      "ReviewActionResponse": {
        "properties": {
          "review_task_id": {
            "type": "string",
            "title": "Review Task Id"
          },
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "review_task_id",
          "target_id",
          "status"
        ],
        "title": "ReviewActionResponse"
      },
      "ReviewTask": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "target_type": {
            "type": "string",
            "title": "Target Type"
          },
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "target_type",
          "target_id",
          "status",
          "payload"
        ],
        "title": "ReviewTask"
      },
...
      "TextImportRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": "default-space"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "TextImportRequest"
      },
      "TextImportResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "chunk_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chunk Ids"
          },
          "candidate_knowledge_unit_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Candidate Knowledge Unit Ids"
          },
          "review_task_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Task Ids"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "source_id",
          "chunk_ids",
          "candidate_knowledge_unit_ids",
          "review_task_ids"
        ],
        "title": "TextImportResponse"
      },
...
    }
  }
}